# Fix memory based performance problem caused by use of lodash memoize in box component