--- a/ui/components/component-library/box/box.tsx
+++ b/ui/components/component-library/box/box.tsx
@@ -130,7 +130,7 @@ const generateClassNames = memoize(
     }
     return classNamesObject;
   },
-  (styleDeclaration, value) => [styleDeclaration, value],
+  (styleDeclaration, value) => `${styleDeclaration}${value}`,
 );
 
 export const Box: BoxComponent = React.forwardRef(